+++ osu.Game/Screens/Select/BeatmapCarousel.cs
@@ -411,7 +411,13 @@ namespace osu.Game.Screens.Select
         public void ScrollToSelected() => scrollPositionCache.Invalidate();
 
         protected override bool OnKeyDown(KeyDownEvent e)
-        {
+        {   
+            // allow for controlling volume when alt is held.
+            // this is required as the VolumeControlReceptor uses OnPressed, which is
+            // executed after all OnKeyDown events.
+            if (e.AltPressed)
+                return base.OnKeyDown(e);
+
             int direction = 0;
             bool skipDifficulties = false;